backend/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
APScheduler==3.10.4
twilio==8.10.2
stripe==7.5.0
razorpay==1.4.1
opencv-python==4.8.1.78
numpy==1.26.2
passlib==1.7.4
bcrypt==4.1.2
